Ensure AuthorizeHttpClientDecorator deletes cache on ForbiddenError

+++ lib/main/decorators/authorize_http_client_decorator.dart
@@ -32,11 +32,13 @@ class AuthorizeHttpClientDecorator implements HttpClient {
         body: body,
         headers: authorizedHeaders,
       );
-    } on HttpError {
-      rethrow;
     } catch (error) {
-      await deleteSecureCacheStorage.deleteSecure('token');
-      throw HttpError.forbbiden;
+      if (error is HttpError && error != HttpError.forbbiden) {
+        rethrow;
+      } else {
+        await deleteSecureCacheStorage.deleteSecure('token');
+        throw HttpError.forbbiden;
+      }
     }
   }
 }
